DO NOT MERGE: filter products from search results
I had to copy the entire search bundle file from jekyll_pages_api_search
and manually run `npm run make-bundle`, but the search is functional.

There's a little bit of a lag once the /products page loads and when it
filters out the non-matching products. Not sure what to do about that yet.

The _real_ solution is to create some more configs/hooks in
jekyll_pages_api_search, not to check this in. Would be nice to get some
frontend tests on it, too.